Add support for plotting per-sd overutilized traces in frequency analysis
When we integrate per-sd overutilization, we will have a slightly
different trace token format, which will have some additional information
embedded to list the CPUs which are in the sched_domain which has just
changed status.

Add support for drawing all overutilized bands which contain the CPU you
are drawing a frequency plot for, attempting to maintain coloring across
both plots when you do a cluster analysis.

Signed-off-by: Chris Redpath <chris.redpath@arm.com>